pyproject.toml
[project]
name = "ibkr-ai-broker"
version = "0.1.0"
description = "IBKR Paper Trading Assistant with LLM + Risk Gate + MCP"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [
    { name = "IBKR AI Broker Team" }
]

dependencies = [
    # Core framework
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    
    # IBKR integration
    "ib-insync>=0.9.86",
    
    # Database & ORM
    "sqlalchemy>=2.0.36",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.10",  # PostgreSQL driver
    "aiosqlite>=0.20.0",  # Async SQLite for dev
    
    # Logging & observability
    "structlog>=24.4.0",
    "python-json-logger>=3.2.1",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.2",
    "httpx>=0.28.1",
    "tenacity>=9.0.0",  # Retry logic
    
    # MCP
    "mcp>=1.1.2",
    
    # Dashboard
    "streamlit>=1.40.0",
    "requests>=2.32.0",
]

[project.optional-dependencies]
dev = [
    # Testing
    "pytest>=8.3.4",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "hypothesis>=6.122.3",  # Property-based testing
    
    # Linting & formatting
    "ruff>=0.8.4",
    
    # Type checking
    "pyright>=1.1.390",
    "mypy>=1.13.0",
    
    # Pre-commit
    "pre-commit>=4.0.1",
    
    # Development utilities
    "ipython>=8.30.0",
    "rich>=13.9.4",  # Pretty printing
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["apps", "packages"]

[tool.pytest.ini_options]
testpaths = ["tests", "packages", "apps"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov-report=term-missing",
]
markers = [
    "integration: Integration tests (deselect with '-m \"not integration\"')",
    "slow: Slow tests (deselect with '-m \"not slow\"')",
    "requires_ibkr: Tests that require IBKR connection",
]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.coverage.run]
source = ["packages", "apps"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/.venv/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
    "@overload",
]

[tool.ruff]
target-version = "py312"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "N",      # pep8-naming
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "SIM",    # flake8-simplify
    "TCH",    # flake8-type-checking
    "PTH",    # flake8-use-pathlib
    "RUF",    # ruff-specific rules
]
ignore = [
    "E501",   # Line too long (handled by formatter)
    "B008",   # Do not perform function call in argument defaults (FastAPI depends)
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # Unused imports in __init__.py
"tests/**/*.py" = ["S101", "PLR2004"]  # Allow assert and magic values in tests

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"

[tool.pyright]
include = ["apps", "packages", "tests"]
exclude = [
    "**/__pycache__",
    "**/node_modules",
    ".venv",
    "venv",
]
pythonVersion = "3.12"
typeCheckingMode = "basic"
reportMissingImports = true
reportMissingTypeStubs = false
reportUnusedImport = true
reportUnusedVariable = true

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "ib_insync.*",
    "mcp.*",
]
ignore_missing_imports = true
